% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section_class.R
\name{as_report_section,struct_class,list-method}
\alias{as_report_section,struct_class,list-method}
\title{Creating Report sections from an object}
\usage{
\S4method{as_report_section}{struct_class,list}(obj, subsection)
}
\arguments{
\item{obj}{A \code{struct_class} object}

\item{subsection}{A list of report sections to include as a subsection for this object.}
}
\value{
A \code{report_section} object with pre-populated slots based on the input object.
}
\description{
This method converts the input \code{struct} object into a report section for use with
other reporting objects such as \code{struct_report} and methods like \code{build_report}.
}
