% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{build_report}
\alias{build_report}
\title{Build struct report}
\usage{
build_report(R, outfile)
}
\arguments{
\item{R}{A struct_report object}

\item{outfile}{A path/filename to save the output to}
}
\value{
A rendered markdown file
}
\description{
Builds a struct report using the input data and writes the output to the specified location.
}
