% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_valid}
\alias{is_valid}
\title{Is model valid for report}
\usage{
is_valid(R, M, ...)
}
\arguments{
\item{R}{A struct_report object}

\item{M}{a struct_model object}

\item{...}{additional inputs}
}
\value{
TRUE or throws an error
}
\description{
Checks if the provided model is valid for a specific report object by comparing the class of the model with the model_name slot of all report sections.
}
