% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_centre_section.R
\docType{class}
\name{mean_centre_section-class}
\alias{mean_centre_section-class}
\alias{.mean_centre_section}
\title{mean_centre section}
\description{
This report section summarises the mean_centre model applied to a dataset.
}
