% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_centre_section.R
\name{mean_centre_section}
\alias{mean_centre_section}
\title{Constructor for mean_centre_section objects}
\usage{
mean_centre_section(
  mode = "data",
  markdown = system.file(package = "structReport", "templates/mean_centre_section.Rmd",
    mustWork = TRUE),
  ...
)
}
\arguments{
\item{mode}{Used to control whether centring is apply to the data, the meta data or both.
Can be any one of "data","sample_meta" or "both". default is "data".}

\item{...}{additional slots and values passed to struct_class}
}
\value{
a report_section object
}
\description{
Creates a new \linkS4class{mean_centre_section} object and populates the slots.
}
