% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_section.R
\name{PCA_section}
\alias{PCA_section}
\title{Constructor for PCA_section objects}
\usage{
PCA_section(
  factor_names,
  markdown = system.file(package = "structReport", "templates/PCA_section.Rmd", mustWork
    = TRUE),
  ...
)
}
\value{
a report_section object#' @include report_section_class.R
}
\description{
Creates a new \linkS4class{PCA_section} object and populates the slots.
}
