% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section_class.R
\docType{class}
\name{report_section-class}
\alias{report_section-class}
\alias{.report_section}
\title{report_section object}
\description{
Struct reports contain parameters for the child sections of a struct report.
}
