% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_report_class.R
\docType{class}
\name{struct_report-class}
\alias{struct_report-class}
\alias{.struct_report}
\title{struct_report object}
\description{
Struct reports contain parameters for populating yaml and knitr settings for rmarkdown reports.
}
