% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_report_class.R
\name{struct_report}
\alias{struct_report}
\alias{[,struct_report,ANY,ANY,ANY-method}
\alias{[<-,struct_report,ANY,ANY,ANY-method}
\alias{length,struct_report-method}
\alias{+,report_section,struct_report-method}
\alias{+,struct_report,report_section-method}
\alias{+,report_section,report_section-method}
\alias{+,struct_report,struct_report-method}
\alias{build_report,struct_report,character-method}
\title{Constructor for struct_report objects}
\usage{
struct_report(
  title,
  author = "Anon",
  format = "html_document",
  toc = TRUE,
  toc_depth = 2,
  sections = list(),
  ...
)

\S4method{[}{struct_report,ANY,ANY,ANY}(x, i)

\S4method{[}{struct_report,ANY,ANY,ANY}(x, i) <- value

\S4method{length}{struct_report}(x)

\S4method{+}{report_section,struct_report}(e1, e2)

\S4method{+}{struct_report,report_section}(e1, e2)

\S4method{+}{report_section,report_section}(e1, e2)

\S4method{+}{struct_report,struct_report}(e1, e2)

\S4method{build_report}{struct_report,character}(R, outfile)
}
\arguments{
\item{title}{the title of the report.}

\item{author}{the author of the report}

\item{format}{the format of the report. One of 'html_document','pdf_document','word_document'. Default = 'html_document'}

\item{sections}{a list of report_sections}
}
\value{
a struct_class object

model for the report section at the given index

struct_report with the model for the section at index i replaced

the number of sections in a struct_report

a struct_report with the additional section appended to the front of
the sequence

a struct_report with the additional section appended to the end of the
sequence

a struct report containing the combined sections

a struct report containing a section for each of the input reports
}
\description{
Creates a new \linkS4class{struct_report} object and populates the slots.
}
\examples{
MS = report_section() + report_section()
MS[2]

MS = report_section() + report_section()
MS[2] = model()

MS = report_section() + report_section()
length(MS) # 2

MS = model_report() + model_report()
M = model_report()
MS = M + MS

MS = model_report() + model_report()
M = model_report()
MS = MS + M

MS = report_section() + report_section()

MS = struct_report() + struct_report()

}
