% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_report_list_class.R
\docType{class}
\name{struct_report_list-class}
\alias{struct_report_list-class}
\alias{.struct_report_list}
\title{struct_report_list object}
\description{
Struct report lists contain parameters for populating yaml and knitr settings for combining multiple rmarkdown reports.
}
