% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_report_list_class.R
\name{struct_report_list}
\alias{struct_report_list}
\alias{+,struct_report,struct_report_list-method}
\alias{+,struct_report_list,struct_report-method}
\alias{+,struct_report_list,struct_report_list-method}
\alias{c,struct_report_list-method}
\title{Constructor for struct_report_list objects}
\usage{
struct_report_list(
  title,
  author = "Anon",
  format = "html_document",
  toc = TRUE,
  toc_depth = 2,
  markdown = system.file(package = "structReport", "templates/struct_report.Rmd",
    mustWork = TRUE),
  sections = list(),
  model_name = NULL,
  ...
)

\S4method{+}{struct_report,struct_report_list}(e1, e2)

\S4method{+}{struct_report_list,struct_report}(e1, e2)

\S4method{+}{struct_report_list,struct_report_list}(e1, e2)

\S4method{c}{struct_report_list}(x, ...)
}
\arguments{
\item{title}{the title of the report.}

\item{author}{the author of the report}

\item{format}{the format of the report. One of 'html_document','pdf_document','word_document'. Default = 'html_document'}

\item{markdown}{the markdown template to use}

\item{sections}{a list of report_sections}

\item{model_name}{the name of the model required for this report section}
}
\value{
a struct_class object

a struct_report_list with the additional section appended to the front

a struct_report with the additional section appended to the end

a struct report list containing the combined sections

a struct report list containing the combined sections
}
\description{
Creates a new \linkS4class{struct_report_list} object and populates the slots.
}
\examples{
MS = struct_report() + struct_report()
M = struct_report()
MS = M + MS

MS = struct_report() + struct_report()
M = struct_report()
MS = MS + M

MS = struct_report_list() + struct_report_list()

MS = c(struct_report(), struct_report())

}
