% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section_class.R
\name{report_section}
\alias{report_section}
\title{Constructor for report_section objects}
\usage{
report_section(
  model_name = NULL,
  markdown = system.file(package = "structReport", "templates/report_section.Rmd",
    mustWork = TRUE)
)
}
\arguments{
\item{markdown}{path and filename of markdown template to use for the section}

\item{models}{the model to use for this report section}
}
\value{
a struct_class object
}
\description{
Creates a new \linkS4class{report_section} object and populates the slots.
}
