% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_report_class.R,
%   R/struct_report_list_class.R
\name{struct_report}
\alias{struct_report}
\alias{[,struct_report,ANY,ANY,ANY-method}
\alias{[<-,struct_report,ANY,ANY,ANY-method}
\alias{length,struct_report-method}
\alias{+,report_section,struct_report-method}
\alias{+,struct_report,report_section-method}
\alias{+,report_section,report_section-method}
\alias{+,struct_report,struct_report-method}
\alias{build_report,struct_report,model_OR_model_seq,DatasetExperiment,character-method}
\alias{build_report,struct_report,iterator,DatasetExperiment,character-method}
\alias{build_report,list,list,DatasetExperiment,character-method}
\alias{is_valid,struct_report,model-method}
\alias{is_valid,struct_report,model_seq-method}
\alias{is_valid,struct_report,iterator-method}
\alias{c,struct_report-method}
\title{Constructor for struct_report objects}
\usage{
struct_report(
  title,
  author = "Anon",
  format = "html_document",
  toc = TRUE,
  toc_depth = 2,
  markdown = system.file(package = "structReport", "templates/struct_report.Rmd",
    mustWork = TRUE),
  sections = list(),
  model_name = NULL,
  ...
)

\S4method{[}{struct_report,ANY,ANY,ANY}(x, i)

\S4method{[}{struct_report,ANY,ANY,ANY}(x, i) <- value

\S4method{length}{struct_report}(x)

\S4method{+}{report_section,struct_report}(e1, e2)

\S4method{+}{struct_report,report_section}(e1, e2)

\S4method{+}{report_section,report_section}(e1, e2)

\S4method{+}{struct_report,struct_report}(e1, e2)

\S4method{build_report}{struct_report,model_OR_model_seq,DatasetExperiment,character}(R, M, D, outfile)

\S4method{build_report}{struct_report,iterator,DatasetExperiment,character}(R, M, D, outfile)

\S4method{build_report}{list,list,DatasetExperiment,character}(R, M, D, outfile)

\S4method{is_valid}{struct_report,model}(R, M)

\S4method{is_valid}{struct_report,model_seq}(R, M)

\S4method{is_valid}{struct_report,iterator}(R, M)

\S4method{c}{struct_report}(x, ...)
}
\arguments{
\item{title}{the title of the report.}

\item{author}{the author of the report}

\item{format}{the format of the report. One of 'html_document','pdf_document','word_document'. Default = 'html_document'}

\item{markdown}{the markdown template to use}

\item{sections}{a list of report_sections}

\item{model_name}{the name of the model required for this report section}
}
\value{
a struct_class object

model for the report section at the given index

struct_report with the model for the section at index i replaced

the number of sections in a struct_report

a struct_report with the additional section appended to the front of
the sequence

a struct_report with the additional section appended to the end of the
sequence

a struct report containing the combined sections

a struct report list containing the combined sections

a struct report list containing the combined sections
}
\description{
Creates a new \linkS4class{struct_report} object and populates the slots.
}
\examples{
MS = report_section() + report_section()
MS[2]

MS = report_section() + report_section()
MS[2] = model()

MS = report_section() + report_section()
length(MS) # 2

MS = model_report() + model_report()
M = model_report()
MS = M + MS

MS = model_report() + model_report()
M = model_report()
MS = MS + M

MS = model_report() + model_report()

MS = struct_report() + struct_report()

MS = c(struct_report(), struct_report())

}
